Refactor
#182

Refactors ScrumblesFrames into its own file in such a way that imports from views do not change.

SFrames as split into SDynamics, SGraphs, and SLists.
