+++ source/ScrumblesFrames/SGraphs.py
@@ -1,0 +1,58 @@
+import tkinter as tk
+import webbrowser
+import matplotlib
+matplotlib.use("TKAgg")
+from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
+from matplotlib.figure import Figure
+
+import csv
+import tkcalendar
+import datetime
+import ScrumblesData
+import ScrumblesObjects
+
+from SDynamics import *
+from SLists import *
+
+from styling import styling as style
+from tkinter import ttk
+
+class SLineGraph(tk.Frame):
+    def __init__(self, controller):
+        tk.Frame.__init__(self, controller)
+
+        self.x = [1,2,3,4,5,6,7,8,9,10]
+        self.y = [1,4,2,5,1,4,2,5,1,4]
+        self.label = None
+
+        self.figure = Figure(figsize=(4, 4), dpi=100)
+        self.graph = self.figure.add_subplot(1,1,1)
+
+
+    def setTitle(self, title):
+        self.label = tk.Label(self, text=title)
+
+    def setAxes(self, xAxis, yAxis):
+        self.graph.set_xlabel(xAxis)
+        self.graph.set_ylabel(yAxis)
+        self.figure.subplots_adjust(left=.15)
+        self.figure.subplots_adjust(bottom=.15)
+
+    def importDataFromCSV(self, fileName, delimeter):
+        with open(fileName, 'r') as file:
+            plots = csv.reader(file, delimeter)
+            for row in plots:
+                self.x.append(int(row[0]))
+                self.y.append(int(row[1]))
+
+    def displayGraph(self):
+        self.graph.plot(self.x, self.y)
+        canvas = FigureCanvasTkAgg(self.figure, self)
+        canvas.show()
+        if self.label is not None:
+            self.label.pack(side=tk.TOP, fill=tk.X)
+
+        canvas.get_tk_widget().pack(side=tk.BOTTOM, fill=tk.BOTH, expand=True, padx=1, pady=1)
+
+    def changeBackgroundColor(self, color):
+        self.graph.set_facecolor(facecolor=color)
